pyproject.toml
[project]
name = "fastapiproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.2.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "flake8>=7.3.0",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.2",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
